Address the comments.
