+++ gcp_variant_transforms/beam_io/vcf_header_io.py
@@ -29,7 +29,7 @@ from apache_beam.io.filesystems import FileSystems
 from apache_beam.io.iobase import Read
 from apache_beam.transforms import PTransform
 
-from gcp_variant_transforms.beam_io import vcf_parser
+from gcp_variant_transforms.beam_io import bgzf
 from gcp_variant_transforms.beam_io import vcfio
 
 
@@ -160,7 +160,7 @@ class VcfHeaderSource(filebasedsource.FileBasedSource):
 
   def open_file(self, file_name):
     if self._compression_type == CompressionTypes.GZIP:
-      return vcf_parser.open_bgzf(file_name)
+      return bgzf.open_bgzf(file_name)
     else:
       return FileSystems.open(file_name,
                               compression_type=self._compression_type)
