+++ gcp_variant_transforms/testing/integration/vcf_to_bq_tests/presubmit_tests/medium_tests/test_splittable_gzip.json
@@ -2,7 +2,7 @@
   {
     "test_name": "test-splittable-gzip",
     "table_name": "test_splittable_gzip",
-    "input_pattern": "gs:///gcp-variant-transforms-testfiles/medium_tests/splittable-gzip/**.bgz",
+    "input_pattern": "gs://gcp-variant-transforms-testfiles/medium_tests/splittable-gzip/**.bgz",
     "runner": "DataflowRunner",
     "assertion_configs": [
       {
